+++ examples/combo_box_table/combo_box_grouped/index.jsx
@@ -34,7 +34,7 @@ export function Example() {
   const [value, setValue] = useState(null);
   const [search, setSearch] = useState(null);
   const filteredOptions = useTokenSearch(search, { options: cats, index });
-  const onExpandListBox = useLayoutListBox(layoutMaxWidth, layoutMaxHeight, layoutColumnsAlignLeft);
+  const onLayoutListBox = useLayoutListBox(layoutMaxWidth, layoutMaxHeight, layoutColumnsAlignLeft);
 
   return (
     <>
@@ -54,7 +54,7 @@ export function Example() {
         columns={columns}
         renderColumnValue={highlightValue(highlighter)}
         mapOption={mapOption}
-        onExpandListBox={onExpandListBox}
+        onLayoutListBox={onLayoutListBox}
       />
 
       <label htmlFor="output">