+++ examples/drop_down_table/drop_down_grouped/index.jsx
@@ -11,7 +11,7 @@ function mapOption({ breed, coatLength }) {
 export function Example() {
   const [value, setValue] = useState(null);
   const ref = useRef();
-  const onExpandListBox = useLayoutListBox(layoutMaxWidth, layoutMaxHeight);
+  const onLayoutListBox = useLayoutListBox(layoutMaxWidth, layoutMaxHeight);
   const [managedFocus, setManagedFocus] = useState(true);
 
   return (
@@ -29,7 +29,7 @@ export function Example() {
         ref={ref}
         value={value}
         onValue={setValue}
-        onExpandListBox={onExpandListBox}
+        onLayoutListBox={onLayoutListBox}
         options={cats}
         columns={columns}
         mapOption={mapOption}